Use yt-dlp library to download media instead of subprocess
Running yt-dlp as a keyword in the command line wasn't as portable and as reliable as I'd hoped, so we're pivoting to using yt_dlp directly as a Python package.

As far as I can tell, this change has also eliminated our reliance on FFmpeg and FFprobe for reliable audio-only media.