Refactor file paths for artifact uploads
The entire pyinstaller file is now uploaded for each platform that the program is compiled on. The executable is still the only thing attached to the release, but this change allows me to download the entire pyinstaller folder from the upload-artifact action and poke around to make sure that everything is working properly. I don't expect this change to fix anything, but it should help me determine the source of the error